add integer to model, solver int as result

+++ solver.cpp
@@ -70,7 +70,7 @@ int main( int argc, char *argv[] )
 			printf("FINAL_LOG:Reqular Tickets:\n");
 		if(i == (columns_count/2))
 			printf("FINAL_LOG:Premium Tickets:\n");
-		printf( "FINAL_LOG:e[%d] = %g\n", (i + 1)%(columns_count/2+1), column_primal[ i ] );
+		printf( "FINAL_LOG:e[%d] = %d\n", (i + 1)%(columns_count/2+1), (int)column_primal[ i ] );
 	}
 //	for ( i = 0; i < columns_count; ++i )
 //	{
